requirements.txt
streamlit==1.29.0
streamlit-file-uploader==0.1.1
Pillow==10.1.0
piexif==1.1.3